# perf: keep using the lightweight Response object when retrieving `headers`, `status`, and `ok`, and then drop the `getInternalBody` function.