This PR adopts https://github.com/usualoma/node-server/pull/3, excludes https://github.com/honojs/node-server/pull/241, and completely removes the functionality of `getInternalBody`.

### support more response body types for caching

In d72de8d, I made the optimization pattern previously covered by `getInternalBody` return from a faster cache.


### benchmark

#### script

```ts
import { serve } from '@hono/node-server'
import { Hono } from 'hono'

const app = new Hono()

app.get('/', () => {
    const res = new Response('foo')
    res.headers.set('x-foo', 'bar')
    return res
})

app.get('/uint8array', () => {
    return new Response(new Uint8Array([102, 111, 111])) // prints "foo"
})

app.get('/blob', () => {
    return new Response(new Blob([new Uint8Array([102, 111, 111])])) // prints "foo"
})

serve({
    fetch: app.fetch,
    port: 3003,
})

```

#### main branch

```
$ bombardier -d 10s --fasthttp http://localhost:3003/
Bombarding http://localhost:3003/ for 10s using 125 connection(s)
[==========================================================================] 10s
Done!
Statistics        Avg      Stdev        Max
  Reqs/sec     41208.28    6649.70   51394.11
  Latency        3.03ms     3.18ms   328.28ms
  HTTP codes:
    1xx - 0, 2xx - 411938, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:     9.39MB/s
$ bombardier -d 10s --fasthttp http://localhost:3003/uint8array
Bombarding http://localhost:3003/uint8array for 10s using 125 connection(s)
[==========================================================================] 10s
Done!
Statistics        Avg      Stdev        Max
  Reqs/sec     39105.54    7149.83   48899.74
  Latency        3.19ms     1.58ms   143.78ms
  HTTP codes:
    1xx - 0, 2xx - 390974, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:     8.87MB/s
$ bombardier -d 10s --fasthttp http://localhost:3003/blob
Bombarding http://localhost:3003/blob for 10s using 125 connection(s)
[==========================================================================] 10s
Done!
Statistics        Avg      Stdev        Max
  Reqs/sec     34814.79    7252.95   48210.01
  Latency        3.59ms     1.25ms    83.12ms
  HTTP codes:
    1xx - 0, 2xx - 348063, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:     7.70MB/s
```

#### usualoma:feat-more-use-respose-cache-2 branch

```
$ bombardier -d 10s --fasthttp http://localhost:3003/
Bombarding http://localhost:3003/ for 10s using 125 connection(s)
[==========================================================================] 10s
Done!
Statistics        Avg      Stdev        Max
  Reqs/sec     50349.82    4485.60   61434.61
  Latency        2.48ms   412.72us    30.42ms
  HTTP codes:
    1xx - 0, 2xx - 503307, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:    11.52MB/s
$ bombardier -d 10s --fasthttp http://localhost:3003/uint8array
Bombarding http://localhost:3003/uint8array for 10s using 125 connection(s)
[==========================================================================] 10s
Done!
Statistics        Avg      Stdev        Max
  Reqs/sec     49739.84    6358.60   59130.66
  Latency        2.51ms     1.68ms   143.02ms
  HTTP codes:
    1xx - 0, 2xx - 497340, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:    11.29MB/s
$ bombardier -d 10s --fasthttp http://localhost:3003/blob
Bombarding http://localhost:3003/blob for 10s using 125 connection(s)
[==========================================================================] 10s
Done!
Statistics        Avg      Stdev        Max
  Reqs/sec     40411.95    6107.91   51111.33
  Latency        3.09ms     2.07ms   178.87ms
  HTTP codes:
    1xx - 0, 2xx - 403772, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:     8.94MB/s
```